+++ src/QueryExpressionFilter.php
@@ -23,8 +23,6 @@ class QueryExpressionFilter implements Filter
     {
         if ($this->isObject($filter)) {
             return $this->matchAnd($data, $filter);
-        } elseif (is_bool($filter)) {
-            return $filter;
         } else {
             throw new DomainException('Invalid filter type');
         }
@@ -32,26 +30,26 @@ class QueryExpressionFilter implements Filter
 
     private function matchOr($data, $filter)
     {
+        $empty = true;
         if ($this->isVector($filter)) {
-            if (!$filter) {
-                return true;
-            }
             foreach ($filter as $element) {
                 if ($this->matchFilter($data, $element)) {
                     return true;
                 }
+                $empty = false;
             }
         } elseif ($this->isObject($filter)) {
             foreach ($filter as $key => $value) {
                 if ($this->matchValue($data, $key, $value)) {
                     return true;
                 }
+                $empty = false;
             }
         } else {
             throw new DomainException('Invalid data type for $or combinator');
         }
 
-        return false;
+        return $empty;
     }
 
     private function matchAnd($data, $filter)
@@ -92,23 +90,21 @@ class QueryExpressionFilter implements Filter
         $expectedValue = $expectation;
         $comparator = '$is';
 
-        if (is_array($expectedValue)) {
-            if ($this->isVector($expectedValue)) {
-                // list of possible values
-                $comparator = '$in';
-            } else {
-                // custom comparator ('>', '<', ...)
-                $comparator = key($expectedValue);
-                $expectedValue = reset($expectedValue);
+        if ($this->isVector($expectedValue)) {
+            // list of possible values
+            $comparator = '$in';
+        } elseif ($this->isObject($expectedValue)) {
+            // custom comparator ('>', '<', ...)
+            $comparator = key($expectedValue);
+            $expectedValue = reset($expectedValue);
 
-                if ($comparator === '$not') {
-                    $comparator = ($this->isVector($expectedValue)) ? '!$in' : '!$is';
-                }
+            if ($comparator === '$not') {
+                $comparator = ($this->isVector($expectedValue)) ? '!$in' : '!$is';
+            }
 
-                if ($comparator[0] === '!') {
-                    $comparator = substr($comparator, 1);
-                    return !$this->matchValue($data, $column, array($comparator => $expectedValue));
-                }
+            if ($comparator[0] === '!') {
+                $comparator = substr($comparator, 1);
+                return !$this->matchValue($data, $column, array($comparator => $expectedValue));
             }
         }
 
@@ -135,13 +131,25 @@ class QueryExpressionFilter implements Filter
 
     private function fetchValue($data, $column)
     {
-        // TODO: dotted notation
-        return isset($data[$column]) ? $data[$column] : null;
+        $path = explode($this->selectorSeparator, $column);
+
+        $current = $data;
+        foreach ($path as $field) {
+            if (is_array($current) && isset($current[$field])) {
+                $current = $current[$field];
+            } elseif (is_object($current) && isset($current->$field)) {
+                $current = $current->$field;
+            } else {
+                return null;
+            }
+        }
+
+        return $current;
     }
 
     private function isObject($value)
     {
-        return (is_array($value) && ($value === array() || !isset($value[0])));
+        return (is_object($value) || (is_array($value) && ($value === array() || !isset($value[0]))));
     }
 
     private function isVector($value)