+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## 0.2.0 (2014-11-04)
+
+* Feature: Support dotted notation for nested attributes
+  (#5 by @maximkou)
+
+* Remove: Remove boolean filters (dropped from spec v0.3)
+  (#8)
+
 ## 0.1.0 (2014-08-21)
 
 * First tagged release